requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
rembg>=2.0.50
torch>=2.0.0
torchvision>=0.15.0
segment-anything>=1.0
opencv-python>=4.8.0
simple-lama-inpainting>=0.1.3